requirements.txt
requests
Django==1.8
djangorestframework
markdown
django-filter
dateparser
